+++ lib/disposable/twin/property/hash.rb
@@ -1,4 +1,4 @@
-require "disposable/twin/struct"
+require "disposable/twin/property/struct"
 
 class Disposable::Twin
   module Property