+++ test/twin/hash_test.rb
@@ -1,5 +1,7 @@
-require "test_helper"
-require "disposable/twin/property/hash"
+# frozen_string_literal: true
+
+require 'test_helper'
+require 'disposable/twin/property/hash'
 
 class HashTest < MiniTest::Spec
   Model = Struct.new(:id, :content)
@@ -27,86 +29,85 @@ class HashTest < MiniTest::Spec
 
   # puts Song.definitions.get(:content)[:nested].definitions.get(:band).inspect
 
-  it "allows reading from existing hash" do
+  it 'allows reading from existing hash' do
     model = Model.new(1, {})
-    expect(model.inspect).must_equal "#<struct HashTest::Model id=1, content={}>"
+    _(model.inspect).must_equal '#<struct HashTest::Model id=1, content={}>'
 
     song = Song.new(model)
-    expect(song.id).must_equal 1
-    expect(song.content.title).must_be_nil
-    expect(song.content.band.name).must_be_nil
-    expect(song.content.band.label.location).must_be_nil
-    expect(song.content.releases).must_equal []
+    _(song.id).must_equal 1
+    _(song.content.title).must_be_nil
+    _(song.content.band.name).must_be_nil
+    _(song.content.band.label.location).must_be_nil
+    _(song.content.releases).must_equal []
 
     # model's hash hasn't changed.
-    expect(model.inspect).must_equal "#<struct HashTest::Model id=1, content={}>"
+    _(model.inspect).must_equal '#<struct HashTest::Model id=1, content={}>'
   end
 
-  it "defaults to hash when value is nil" do
+  it 'defaults to hash when value is nil' do
     model = Model.new(1)
-    expect(model.inspect).must_equal "#<struct HashTest::Model id=1, content=nil>"
+    _(model.inspect).must_equal '#<struct HashTest::Model id=1, content=nil>'
 
     song = Song.new(model)
-    expect(song.id).must_equal 1
-    expect(song.content.title).must_be_nil
-    expect(song.content.band.name).must_be_nil
-    expect(song.content.band.label.location).must_be_nil
+    _(song.id).must_equal 1
+    _(song.content.title).must_be_nil
+    _(song.content.band.name).must_be_nil
+    _(song.content.band.label.location).must_be_nil
 
     # model's hash hasn't changed.
-    expect(model.inspect).must_equal "#<struct HashTest::Model id=1, content=nil>"
+    _(model.inspect).must_equal '#<struct HashTest::Model id=1, content=nil>'
   end
 
-  it "#sync writes to model" do
+  it '#sync writes to model' do
     model = Model.new
 
     song = Song.new(model)
-    song.content.band.label.location = "San Francisco"
+    song.content.band.label.location = 'San Francisco'
 
     song.sync
 
-    expect(model.inspect).must_equal "#<struct HashTest::Model id=nil, content={\"band\"=>{\"label\"=>{\"location\"=>\"San Francisco\"}}, \"releases\"=>[]}>"
+    _(model.inspect).must_equal '#<struct HashTest::Model id=nil, content={"band"=>{"label"=>{"location"=>"San Francisco"}}, "releases"=>[]}>'
   end
 
-  it "#appends to collections" do
+  it '#appends to collections' do
     model = Model.new
 
     song = Song.new(model)
     # song.content.releases.append(version: 1) # FIXME: yes, this happens!
-    song.content.releases.append("version" => 1)
+    song.content.releases.append('version' => 1)
 
     song.sync
 
-    expect(model.inspect).must_equal "#<struct HashTest::Model id=nil, content={\"band\"=>{\"label\"=>{}}, \"releases\"=>[{\"version\"=>1}]}>"
+    _(model.inspect).must_equal '#<struct HashTest::Model id=nil, content={"band"=>{"label"=>{}}, "releases"=>[{"version"=>1}]}>'
   end
 
   it "doesn't erase existing, undeclared content" do
-    model = Model.new(nil, {"artist"=>{}})
+    model = Model.new(nil, { 'artist' => {} })
 
     song = Song.new(model)
-    song.content.band.label.location = "San Francisco"
+    song.content.band.label.location = 'San Francisco'
 
     # puts song.content.class.ancestors
     song.sync
 
-    expect(model.inspect).must_equal "#<struct HashTest::Model id=nil, content={\"artist\"=>{}, \"band\"=>{\"label\"=>{\"location\"=>\"San Francisco\"}}, \"releases\"=>[]}>"
+    _(model.inspect).must_equal '#<struct HashTest::Model id=nil, content={"artist"=>{}, "band"=>{"label"=>{"location"=>"San Francisco"}}, "releases"=>[]}>'
   end
 
   it "doesn't erase existing, undeclared content in existing content" do
-    model = Model.new(nil, {"band"=>{ "label" => { "owner" => "Brett Gurewitz" }, "genre" => "Punkrock" }})
+    model = Model.new(nil, { 'band' => { 'label' => { 'owner' => 'Brett Gurewitz' }, 'genre' => 'Punkrock' } })
 
     song = Song.new(model)
-    song.content.band.label.location = "San Francisco"
+    song.content.band.label.location = 'San Francisco'
 
     song.sync
 
-    expect(model.inspect).must_equal "#<struct HashTest::Model id=nil, content={\"band\"=>{\"label\"=>{\"owner\"=>\"Brett Gurewitz\", \"location\"=>\"San Francisco\"}, \"genre\"=>\"Punkrock\"}, \"releases\"=>[]}>"
+    _(model.inspect).must_equal '#<struct HashTest::Model id=nil, content={"band"=>{"label"=>{"owner"=>"Brett Gurewitz", "location"=>"San Francisco"}, "genre"=>"Punkrock"}, "releases"=>[]}>'
   end
 
-
-  describe "features propagation" do
+  describe 'features propagation' do
     module UUID
       def uuid
-        "1224"
+        '1224'
       end
     end
 
@@ -123,21 +124,21 @@ class HashTest < MiniTest::Spec
       end
     end
 
-    it "includes features into all nested twins" do
+    it 'includes features into all nested twins' do
       song = Hit.new(Model.new)
-      expect(song.uuid).must_equal "1224"
-      expect(song.content.uuid).must_equal "1224"
-      expect(song.content.band.uuid).must_equal "1224"
+      _(song.uuid).must_equal '1224'
+      _(song.content.uuid).must_equal '1224'
+      _(song.content.band.uuid).must_equal '1224'
     end
   end
 
-  describe "coercion" do
-    require "disposable/twin/coercion"
+  describe 'coercion' do
+    require 'disposable/twin/coercion'
     class Coercing < Disposable::Twin
       include Property::Hash
       feature Coercion
 
-      property :id, type: const_get("Types::Coercible::#{DRY_TYPES_INT_CONSTANT}")
+      property :id, type: Types::Params::Integer
       property :content, field: :hash do
         property :title
         property :band do
@@ -146,16 +147,16 @@ class HashTest < MiniTest::Spec
       end
     end
 
-    it "coerces" do
+    it 'coerces' do
       song = Coercing.new(Model.new(1))
-      song.id = "9"
-      expect(song.id).must_equal 9
+      song.id = '9'
+      _(song.id).must_equal 9
       song.content.band.name = 18
-      expect(song.content.band.name).must_equal "18"
+      _(song.content.band.name).must_equal '18'
     end
   end
 
-  describe "::unnest" do
+  describe '::unnest' do
     class Unnesting < Disposable::Twin
       feature Sync
       include Property::Hash
@@ -185,23 +186,23 @@ class HashTest < MiniTest::Spec
       # end
     end
 
-    it "exposes reader and writer" do
-      model = Model.new(1, {title: "Bedroom Eyes"})
+    it 'exposes reader and writer' do
+      model = Model.new(1, { title: 'Bedroom Eyes' })
       song = Unnesting.new(model)
 
       # singular scalar accessors
-      expect(song.content.title).must_equal "Bedroom Eyes"
-      expect(song.title).must_equal "Bedroom Eyes"
+      _(song.content.title).must_equal 'Bedroom Eyes'
+      _(song.title).must_equal 'Bedroom Eyes'
 
-      song.title = "Notorious"
-      expect(song.title).must_equal "Notorious"
-      expect(song.content.title).must_equal "Notorious"
+      song.title = 'Notorious'
+      _(song.title).must_equal 'Notorious'
+      _(song.content.title).must_equal 'Notorious'
 
       # singular nested accessors
-      expect(song.band.name).must_be_nil
-      expect(song.content.band.name).must_be_nil
-      song.band.name = "Duran Duran"
-      expect(song.band.name).must_equal "Duran Duran"
+      _(song.band.name).must_be_nil
+      _(song.content.band.name).must_be_nil
+      song.band.name = 'Duran Duran'
+      _(song.band.name).must_equal 'Duran Duran'
     end
   end
 
@@ -228,31 +229,31 @@ class HashTest < MiniTest::Spec
       end
     end
 
-    it "exposes reader and writer" do
+    it 'exposes reader and writer' do
       model = AlbumModel.new(1, [
-        { title: "Sherry", band: { name: 'The Four Seasons', label: { location: 'US' } }, featured_artists: [{ name: 'Frankie Valli' }, { name: 'The Variatones' }] },
-        { title: "Walk Like a Man", band: { name: 'The Four Seasons', label: { location: 'US' } }, featured_artists: [{ name: 'Frankie Valli' }] }
-      ])
+                               { title: 'Sherry', band: { name: 'The Four Seasons', label: { location: 'US' } }, featured_artists: [{ name: 'Frankie Valli' }, { name: 'The Variatones' }] },
+                               { title: 'Walk Like a Man', band: { name: 'The Four Seasons', label: { location: 'US' } }, featured_artists: [{ name: 'Frankie Valli' }] }
+                             ])
       contract = Album.new(model)
 
       song1 = contract.songs[0]
 
-      expect(song1.title).must_equal "Sherry"
-      expect(song1.band.name).must_equal 'The Four Seasons'
-      expect(song1.band.label.location).must_equal 'US'
-      expect(song1.featured_artists[0].name).must_equal 'Frankie Valli'
-      expect(song1.featured_artists[1].name).must_equal 'The Variatones'
+      _(song1.title).must_equal 'Sherry'
+      _(song1.band.name).must_equal 'The Four Seasons'
+      _(song1.band.label.location).must_equal 'US'
+      _(song1.featured_artists[0].name).must_equal 'Frankie Valli'
+      _(song1.featured_artists[1].name).must_equal 'The Variatones'
 
       song2 = contract.songs[1]
 
-      expect(song2.title).must_equal "Walk Like a Man"
-      expect(song2.band.name).must_equal 'The Four Seasons'
-      expect(song2.band.label.location).must_equal 'US'
-      expect(song2.featured_artists[0].name).must_equal 'Frankie Valli'
+      _(song2.title).must_equal 'Walk Like a Man'
+      _(song2.band.name).must_equal 'The Four Seasons'
+      _(song2.band.label.location).must_equal 'US'
+      _(song2.featured_artists[0].name).must_equal 'Frankie Valli'
     end
   end
 end
 
-# fixme: make sure default hash is different for every invocation, and not created at compile time.
+# FIXME: make sure default hash is different for every invocation, and not created at compile time.
 
 # TODO: test that config is same and nested.