+++ test/twin/property_processor_test.rb
@@ -1,9 +1,11 @@
-require "test_helper"
+# frozen_string_literal: true
+
+require 'test_helper'
 
 class PropertyProcessorTest < Minitest::Spec
-	Album  = Struct.new(:title, :artist, :songs)
-	Artist = Struct.new(:name)
-	Song   = Struct.new(:id)
+  Album  = Struct.new(:title, :artist, :songs)
+  Artist = Struct.new(:name)
+  Song   = Struct.new(:id)
 
   class AlbumTwin < Disposable::Twin
     property :title
@@ -13,33 +15,33 @@ class PropertyProcessorTest < Minitest::Spec
     end
 
     collection :songs do
-    	property :id
+      property :id
     end
   end
 
-  describe "collection" do
-  	let(:twin) { AlbumTwin.new(Album.new("Live!", Artist.new, [Song.new(1), Song.new(2)])) }
-  	it "yields twin, index" do
-	    called = []
-	    Disposable::Twin::PropertyProcessor.new(twin.class.definitions.get(:songs), twin).() { |v, i| called << [v.model, i] }
+  describe 'collection' do
+    let(:twin) { AlbumTwin.new(Album.new('Live!', Artist.new, [Song.new(1), Song.new(2)])) }
+    it 'yields twin, index' do
+      called = []
+      Disposable::Twin::PropertyProcessor.new(twin.class.definitions.get(:songs), twin).call { |v, i| called << [v.model, i] }
 
-	    expect(called.inspect).must_equal %{[[#<struct PropertyProcessorTest::Song id=1>, 0], [#<struct PropertyProcessorTest::Song id=2>, 1]]}
-  	end
+      _(called.inspect).must_equal %([[#<struct PropertyProcessorTest::Song id=1>, 0], [#<struct PropertyProcessorTest::Song id=2>, 1]])
+    end
 
-  	it "yields twin" do
-	    called = []
-	    Disposable::Twin::PropertyProcessor.new(twin.class.definitions.get(:songs), twin).() { |v| called << [v.model] }
+    it 'yields twin' do
+      called = []
+      Disposable::Twin::PropertyProcessor.new(twin.class.definitions.get(:songs), twin).call { |v| called << [v.model] }
 
-	    expect(called.inspect).must_equal %{[[#<struct PropertyProcessorTest::Song id=1>], [#<struct PropertyProcessorTest::Song id=2>]]}
-  	end
+      _(called.inspect).must_equal %([[#<struct PropertyProcessorTest::Song id=1>], [#<struct PropertyProcessorTest::Song id=2>]])
+    end
 
-  	it "allows nil collection" do
-  		twin   = AlbumTwin.new(Album.new("Live!", Artist.new, nil))
+    it 'allows nil collection' do
+      twin = AlbumTwin.new(Album.new('Live!', Artist.new, nil))
 
-	    called = []
-	    Disposable::Twin::PropertyProcessor.new(twin.class.definitions.get(:songs), twin).() { |v, i| called << [v.model, i] }
+      called = []
+      Disposable::Twin::PropertyProcessor.new(twin.class.definitions.get(:songs), twin).call { |v, i| called << [v.model, i] }
 
-	    expect(called.inspect).must_equal %{[]}
-  	end
+      _(called.inspect).must_equal %([])
+    end
   end
 end