+++ test/twin/virtual_test.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'test_helper'
 
 class VirtualTest < MiniTest::Spec
@@ -6,12 +8,12 @@ class VirtualTest < MiniTest::Spec
     property :credit_card_number, virtual: true # no read, no write, it's virtual.
   end
 
-  let (:twin) { CreditCardTwin.new(Object.new) }
+  let(:twin) { CreditCardTwin.new(Object.new) }
 
   it {
-    twin.credit_card_number = "123"
+    twin.credit_card_number = '123'
 
-    expect(twin.credit_card_number).must_equal "123"  # this is still readable in the UI.
+    _(twin.credit_card_number).must_equal '123' # this is still readable in the UI.
 
     twin.sync
 
@@ -20,20 +22,20 @@ class VirtualTest < MiniTest::Spec
       hash = nested
     end
 
-    expect(hash).must_equal("credit_card_number"=> "123")
+    _(hash).must_equal('credit_card_number' => '123')
   }
 
-  describe "setter should never be called with virtual:true" do
+  describe 'setter should never be called with virtual:true' do
     class Raising < Disposable::Twin
       property :id, virtual: true
 
       def id=(*)
-        raise "i should never be called!"
+        raise 'i should never be called!'
       end
     end
 
-    it "what" do
-      expect(Raising.new(Object.new).id).must_be_nil
+    it 'what' do
+      _(Raising.new(Object.new).id).must_be_nil
     end
   end
 end