+++ test/twin/writeable_test.rb
@@ -1,14 +1,16 @@
+# frozen_string_literal: true
+
 require 'test_helper'
 
 class WriteableTest < MiniTest::Spec
   Credentials = Struct.new(:password, :credit_card) do
-    def password=(v)
+    def password=(_v)
       raise "don't call me!"
     end
   end
 
   CreditCard = Struct.new(:name, :number) do
-    def number=(v)
+    def number=(_v)
       raise "don't call me!"
     end
   end
@@ -25,25 +27,25 @@ class WriteableTest < MiniTest::Spec
     end
   end
 
-  let (:cred) { Credentials.new("secret", CreditCard.new("Jonny", "0987654321")) }
+  let(:cred) { Credentials.new('secret', CreditCard.new('Jonny', '0987654321')) }
 
-  let (:twin) { PasswordForm.new(cred) }
+  let(:twin) { PasswordForm.new(cred) }
 
   it {
-    expect(twin.password).must_equal "secret"
-    expect(twin.credit_card.name).must_equal "Jonny"
-    expect(twin.credit_card.number).must_equal "0987654321"
+    _(twin.password).must_equal 'secret'
+    _(twin.credit_card.name).must_equal 'Jonny'
+    _(twin.credit_card.number).must_equal '0987654321'
 
     # manual setting on the twin works.
-    twin.password = "123"
-    expect(twin.password).must_equal "123"
+    twin.password = '123'
+    _(twin.password).must_equal '123'
 
-    twin.credit_card.number = "456"
-    expect(twin.credit_card.number).must_equal "456"
+    twin.credit_card.number = '456'
+    _(twin.credit_card.number).must_equal '456'
 
     twin.sync
 
-    expect(cred.inspect).must_equal '#<struct WriteableTest::Credentials password="secret", credit_card=#<struct WriteableTest::CreditCard name="Jonny", number="0987654321">>'
+    _(cred.inspect).must_equal '#<struct WriteableTest::Credentials password="secret", credit_card=#<struct WriteableTest::CreditCard name="Jonny", number="0987654321">>'
 
     # test sync{}.
     hash = {}
@@ -51,6 +53,6 @@ class WriteableTest < MiniTest::Spec
       hash = nested
     end
 
-    expect(hash).must_equal("password"=> "123", "credit_card"=>{"name"=>"Jonny", "number"=>"456"})
+    _(hash).must_equal('password' => '123', 'credit_card' => { 'name' => 'Jonny', 'number' => '456' })
   }
 end